+++ lib/helper.js
@@ -129,7 +129,7 @@ const activeLineup = function activeLineup(lineup, dealer, state, rc) {
  * @param
  * @returns
  */
-const wasInvolved = function wasInvolved(lineup, pos, state, rc) {
+const wasInvolved = function wasInvolved(lineup, pos, state) {
   if (typeof pos === 'undefined' || !lineup) {
     throw new Error(`bad param pos: ${pos}`);
   }
@@ -301,11 +301,11 @@ PokerHelper.prototype.inLineup = function inLineup(signer, lineup) {
  */
 PokerHelper.prototype.getMyMaxBet = function getMyMaxBet(lineup, address) {
   if (!lineup || !address) {
-    return -1;
+    throw new Error('invalid params.');
   }
   const myPos = this.getMyPos(lineup, address);
   if (myPos < 0) {
-    return -1;
+    throw new Error(`address ${address} not in lineup.`);
   }
   const receipt = (lineup[myPos].last) ? this.rc.get(lineup[myPos].last) : undefined;
   return (receipt) ? receipt.values[1] : 0;
@@ -338,12 +338,12 @@ PokerHelper.prototype.getMyPos = function getMyPos(lineup, address) {
  */
 PokerHelper.prototype.findMinRaiseAmount = function findMinRaiseAmount(lineup, dealer, lastRoundMaxBet, state) {
   if (!lineup || dealer === undefined) {
-    return -1;
+    throw new Error('invalid params.');
   }
   // TODO: pass correct state
   const max = this.getMaxBet(lineup, state);
   if (max.amount === lastRoundMaxBet) {
-    return -1;
+    throw new Error('can not find minRaiseAmount.');
   }
   // finding second highest bet
   let secondMax = lastRoundMaxBet;
@@ -376,6 +376,9 @@ PokerHelper.prototype.getMaxBet = function getMaxBet(lineup, state) {
       max = amount;
       pos = i;
     }
+  }
+  if (pos === -1) {
+    throw new Error('could not find max bet.');
   }
   return { amount: max, pos };
 };
@@ -417,11 +420,14 @@ PokerHelper.prototype.getWhosTurn = function getWhosTurn(lineup, dealer, state, 
     throw new Error('no ones turn if hand complete.');
   }
   const activePlayers = activeLineup(lineup, dealer, state, this.rc);
+  let maxBet;
 
-  const maxBet = this.getMaxBet(activePlayers, state);
-  if (!evenLineup(activePlayers, this.rc)) {
-    // if betting amounts are not even return next player of to the maxBettor
-    return activePlayers[(maxBet.pos + 1) % activePlayers.length].pos;
+  if (state !== 'waiting') {
+    maxBet = this.getMaxBet(activePlayers, state);
+    if (!evenLineup(activePlayers, this.rc)) {
+      // if betting amounts are not even return next player of to the maxBettor
+      return activePlayers[(maxBet.pos + 1) % activePlayers.length].pos;
+    }
   }
   const checker = findLatest({ sorted: true, lineup: activePlayers }, `check${state}`, this.rc);
   if (checker.last) {
@@ -463,7 +469,7 @@ PokerHelper.prototype.nextPlayer = function nextPlayer(lineup, startPos, type, s
         break;
       }
     } else {
-      newPos = -1;
+      throw new Error('invalid state.');
     }
   }
   return newPos;
@@ -502,15 +508,15 @@ PokerHelper.prototype.getSbPos = function getSbPos(lineup, dealer, state) {
  */
 PokerHelper.prototype.getBbPos = function getBbPos(lineup, dealer, state) {
   if (typeof dealer === 'undefined' || !lineup) {
-    return -1;
+    throw new Error('invalid params.');
   }
   if (dealer < 0 || dealer >= lineup.length || lineup.length < 2) {
-    return -1;
+    throw new Error('invalid more params.');
   }
 
   const next = this.nextPlayer(lineup, dealer + 1, 'involved', state);
   if (next === dealer) {
-    return -1;
+    throw new Error('only one player found.');
   }
   const activePlayerCount = countPlayers(lineup, 'involved', state, this.rc);
   if (activePlayerCount === 2) {
@@ -518,7 +524,7 @@ PokerHelper.prototype.getBbPos = function getBbPos(lineup, dealer, state) {
   }
   const nextNext = this.nextPlayer(lineup, next + 1, 'involved', state);
   if (nextNext === next) {
-    return -1;
+    throw new Error('only one player found.');
   }
   return nextNext;
 };
@@ -580,8 +586,15 @@ PokerHelper.prototype.isBettingDone = function isBettingDone(lineup, dealer, sta
     return true;
   }
   const activePlayers = activeLineup(lineup, dealer, state, this.rc);
-
-  const maxBet = this.getMaxBet(activePlayers, state);
+  let maxBet;
+  try {
+    maxBet = this.getMaxBet(activePlayers, state);
+  } catch (err) {
+    if (state === 'waiting') {
+      return false;
+    }
+    throw err;
+  }
   if (state === 'waiting') {
     return maxBet.amount > 0;
   }
@@ -595,9 +608,9 @@ PokerHelper.prototype.isBettingDone = function isBettingDone(lineup, dealer, sta
     return false;
   }
 
-  let checker = findLatest({ sorted: true, lineup: activePlayers }, 'check', this.rc);
+  const checker = findLatest({ sorted: true, lineup: activePlayers }, 'check', this.rc);
   if (checker.last) {
-    const lastRec = this.rc.get(lineup[activePlayers[activePlayers.length-1].pos].last);
+    const lastRec = this.rc.get(lineup[activePlayers[activePlayers.length - 1].pos].last);
     const checkRec = this.rc.get(lineup[checker.pos].last);
     if (lastRec.abi[0].name.toLowerCase() === `check${state}` &&
       checkRec.abi[0].name.toLowerCase() === `check${state}`) {
@@ -778,7 +791,7 @@ PokerHelper.prototype.calcDistribution = function calcDistribution(lineup, state
   for (i = 0; i < evals.length; i += 1) {
     for (j = 0; j < players.length; j += 1) {
       if (players[j].amount > 0) {
-        let limit = (i > 0) ? evals[i].limit - evals[i-1].limit : evals[i].limit;
+        const limit = (i > 0) ? evals[i].limit - evals[i - 1].limit : evals[i].limit;
         let contribution = limit;
         if (limit > players[j].amount) {
           contribution = players[j].amount;
