+++ lib/helper.spec.js
@@ -1,6 +1,5 @@
 import chai, { expect } from 'chai';
 import sinonChai from 'sinon-chai';
-import Solver from 'pokersolver';
 import EWT from 'ethereum-web-token';
 import PokerHelper from './helper';
 
@@ -12,10 +11,6 @@ const ABI_FOLD = [{ name: 'fold', type: 'function', inputs: [{ type: 'uint' }, {
 const ABI_SIT_OUT = [{ name: 'sitOut', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
 
 const ABI_CHECK_PRE = [{ name: 'checkPre', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
-const ABI_CHECK_FLOP = [{ name: 'checkFlop', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
-const ABI_CHECK_TURN = [{ name: 'checkTurn', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
-const ABI_CHECK_RIVER = [{ name: 'checkRiver', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
-
 const ABI_SHOW = [{ name: 'show', type: 'function', inputs: [{ type: 'uint' }, { type: 'uint' }] }];
 
 // secretSeed: 'rural tent test net drip fatigue uncle action repeat couple lawn rival'
@@ -37,757 +32,6 @@ const P4_PRIV = '0xa803ed744543e69b5e4816c5fc7539427a2928e78d729c87712f180fae52f
 const P_EMPTY = '0x0000000000000000000000000000000000000000';
 
 
-describe('getWhosTurn() in state Dealing ', () => {
-  it('should find BB in turn when SB.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(0);
-  });
-
-  it('should find BB when SB completes preflop.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.equal(2);
-  });
-
-  it('should find BB when SB completes preflop headsup.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.equal(2);
-  });
-
-  it('should throw error with undefined BB.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn.bind(pokerHelper, hand.lineup, hand.dealer, hand.state, undefined)).to.throw('BB amount cannot be undefined');
-  });
-
-  it('should throw Error if heads up and one sitout.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        sitout: 1,
-      }, {
-        address: P2_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn.bind(pokerHelper, hand.lineup, hand.dealer, hand.state)).to.throw('no ones turn if hand complete.');
-  });
-
-  it('should throw Error if hand state is waiting and not enough players.', () => {
-    const hand = {
-      state: 'waiting',
-      lineup: [{
-        address: P1_ADDR,
-      }, {
-        address: P_EMPTY,
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn.bind(pokerHelper, hand.lineup, hand.dealer, hand.state)).to.throw('no ones turn if hand complete.');
-  });
-
-  it('should return small blind if hand state is waiting.', () => {
-    const hand = {
-      state: 'waiting',
-      lineup: [{
-        address: P1_ADDR,
-      }, {
-        address: P2_ADDR, // SB
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find turn of SB in state waiting headsup with sitouter.', () => {
-    const hand = {
-      dealer: 1,
-      state: 'waiting',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 0).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR, // SB
-      }, {
-        address: P_EMPTY,
-      }],
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find turn of SB in state waiting with sitouter.', () => {
-    const hand = {
-      dealer: 1,
-      state: 'waiting',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 0).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR, // SB
-      }],
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(2);
-  });
-
-  it('should identify pos of button.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(1);
-  });
-
-  it('should find turn when not all 0 receipts posted.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 0).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(1);
-  });
-
-  it('should find turn with empty seats.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P4_ADDR,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(3);
-  });
-
-  it('should find turn when not all 0 receipts posted with sitout.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 0).sign(P4_PRIV),
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR,
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P4_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(1);
-  });
-
-  it('should find turn when player timeouted.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        sitout: 1,
-      }, {
-        address: P2_ADDR,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(1);
-  });
-
-
-  // TODO why is this needed?
-  it('should find turn of BB when state dealing.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.eql(1);
-  });
-
-  // TODO why is this needed?
-  it('should identify pos of BB when state dealing with array wrap around.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(0);
-  });
-});
-
-describe('whosTurn() in state PreFlop ', () => {
-  it('should find turn of SB.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.eql(0);
-  });
-
-  it('should find turn of SB with array wrap around.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR, // BB
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR, // SB
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.eql(2);
-  });
-
-  it('should find turn of UTG+1.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 0).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 25)).to.eql(1);
-  });
-
-  it('should find turn when there are sitouts.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P1_ADDR,
-        sitout: 1,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 1).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 0).sign(P4_PRIV),
-        sitout: 1,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.eql(0);
-  });
-
-  it('should find turn when there are timeouts.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P1_ADDR,
-        sitout: 1,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 1).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P4_PRIV),
-        sitout: 1,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state, 50)).to.eql(0);
-  });
-});
-
-describe('whosTurn() in state Flop ', () => {
-  it('should find turn with button checkPre heads up.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_CHECK_PRE).checkPre(1, 100).sign(P3_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }],
-      dealer: 4,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find turn with button checkPre.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_CHECK_PRE).checkPre(1, 100).sign(P3_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }],
-      dealer: 3,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(4);
-  });
-
-  it('should find turn of UTG+1 in flop.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(2);
-  });
-
-  it('should find turn of BB.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 0).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,  // SB
-        last: new EWT(ABI_BET).bet(1, 150).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,  // BB
-        last: new EWT(ABI_BET).bet(1, 102).sign(P3_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 102).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(2);
-  });
-
-  it('should find turn in CO to act on the flop when BB as first to act, checked the flop.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 50).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 100).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-      }],
-      dealer: 3,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(3);
-  });
-
-  it('should find turn of Dealer(1) after raise.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find turn of SB.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P3_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(2);
-  });
-
-  it('should find turn after last check.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 10).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 10).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 10).sign(P3_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(3);
-  });
-
-  it('should find turn next after folded player.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,  // BB
-        last: new EWT(ABI_FOLD).fold(1, 10).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 20).sign(P2_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 20).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR, // SB
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 20).sign(P4_PRIV),
-      }],
-      dealer: 3,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find turn when last to act folded.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 10).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 20).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 20).sign(P3_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find last active to call all in on turn.', () => {
-    expect(pokerHelper.getWhosTurn([{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 500).sign(P1_PRIV),
-    }, {
-      address: P_EMPTY,
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 1000).sign(P3_PRIV),
-      sitout: 'allin',
-    }],
-    0, // dealer
-    'flop')).to.equal(0);
-  });
-
-  it('should find turn after uncalled bet.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 5).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 20).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 5).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 10).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(3);
-  });
-
-  it('should not find turn of folded player.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 5).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 10).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 5).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 10).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(3);
-  });
-});
-
-describe('whosTurn() in state Turn ', () => {
-  it('should distinguish between check rounds.', () => {
-    const hand = {
-      state: 'turn',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 20).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 20).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 20).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 20).sign(P4_PRIV),
-      }],
-      dealer: 2,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(0);
-  });
-});
-
-
-describe('whosTurn() in state River ', () => {
-  it('should find SB in turn after checks in Turn.', () => {
-    const hand = {
-      state: 'river',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 0).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(1);
-  });
-
-  it('should find SB in turn after checkn in Turn (with wraparound).]', () => {
-    const hand = {
-      state: 'river',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 100).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 3,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.equal(0);
-  });
-});
-
-describe('whosTurn() in state Showdown ', () => {
-  it('should find turn of allin in showdown.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-        sitout: 1,
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-        sitout: 'allin',
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 2,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(2);
-  });
-
-  it('should find last to show.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-        sitout: 'allin',
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 100).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.getWhosTurn(hand.lineup, hand.dealer, hand.state)).to.eql(1);
-  });
-});
-
 describe('nextActivePlayer() ', () => {
   it('should be tested');
   // , () => {
@@ -807,427 +51,6 @@ describe('nextActivePlayer() ', () => {
   //     deck: deck
   //   }]});
   // });
-});
-
-describe('isHandComplete() ', () => {
-  it('should reject if not all active players have shown their hand.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 0).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(false);
-  });
-
-  it('should return true with sitout.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_CHECK_RIVER).checkRiver(1, 40).sign(P2_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should allow next hand if all active players have shown their hand.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 1).sign(P4_PRIV),
-        sitout: 1,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should allow next hand if player timed out.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        sitout: 1,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should prevent next hand if all-in player didn\'t show.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        sitout: 'allin',
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(false);
-  });
-
-  it('should allow next hand if all but one player have folded.', () => {
-    const hand = {
-      state: 'flop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 0).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should allow next hand if all but one player have folded or are on sitOut.', () => {
-    const hand = {
-      state: 'dealing',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 40).sign(P1_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_SIT_OUT).sitOut(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 0).sign(P4_PRIV),
-      }, {
-        address: P_EMPTY,
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should allow next hand if all but one player have folded or are timed out.', () => {
-    const hand = {
-      state: 'preflop',
-      lineup: [{
-        address: P1_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 40).sign(P1_PRIV),
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-        last: new EWT(ABI_BET).bet(1, 40).sign(P3_PRIV),
-        sitout: 1,
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_BET).bet(1, 0).sign(P4_PRIV),
-      }],
-      dealer: 0,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-
-  it('should allow next hand if player folded in showdown.', () => {
-    const hand = {
-      state: 'showdown',
-      lineup: [{
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 30).sign(P2_PRIV),
-      }, {
-        address: P3_ADDR,
-          // last: new EWT(ABI_BET).bet(1, 40).sign(P3_PRIV) <- why is this not complet?
-        last: new EWT(ABI_SHOW).show(1, 40).sign(P3_PRIV),
-      }, {
-        address: P4_ADDR,
-        last: new EWT(ABI_FOLD).fold(1, 40).sign(P4_PRIV),
-      }],
-      dealer: 1,
-    };
-    expect(pokerHelper.isHandComplete(hand.lineup, hand.dealer, hand.state)).to.equal(true);
-  });
-});
-
-describe('isBettingDone() ', () => {
-  it('should keep waiting state till SB payed.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-    }, {
-      address: P_EMPTY,
-    }, {
-      address: P3_ADDR,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'waiting', 50)).to.equal(false);
-  });
-
-  it('should proceed to dealing state when SB payed.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 50).sign(P1_PRIV),
-    }, {
-      address: P_EMPTY,
-    }, {
-      address: P3_ADDR,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'waiting', 50)).to.equal(true);
-  });
-
-  it('should proceed to preflop state when 0R payed.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 0).sign(P1_PRIV),
-    }, {
-      address: P_EMPTY,
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'dealing', 50)).to.equal(true);
-  });
-
-  it('should proceed to next round when bb folded.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_FOLD).fold(1, 200).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 200)).to.equal(true);
-  });
-
-  it('should proceed to flop whenn bb checked.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_PRE).checkPre(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      sitout: 1,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 200)).to.equal(true);
-  });
-
-  it('should keep preflop until BB.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 200)).to.equal(false);
-  });
-
-  it('should proceed to next round with folded player.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_PRE).checkPre(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_FOLD).fold(1, 5).sign(P3_PRIV),
-    }, {
-      address: P_EMPTY,
-    }, {
-      address: P_EMPTY,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 2, 'preflop', 200)).to.equal(true);
-  });
-
-  it('should complete preflop when bb timeoutet', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-      sitout: 1,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 100)).to.equal(true);
-  });
-
-  it('should take sitouts into account.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-      sitout: 1,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 100)).to.equal(true);
-  });
-
-  it('should take empty seats into account.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 200).sign(P2_PRIV),
-    }, {
-      address: P_EMPTY,
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'preflop', 100)).to.equal(true);
-  });
-
-  it('should wait for all to post 0 receipt.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_BET).bet(1, 50).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
-    }, {
-      address: P4_ADDR,
-      last: new EWT(ABI_BET).bet(1, 0).sign(P4_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'dealing')).to.equal(false);
-  });
-
-  it('should wait for all to check.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 150).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 150).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_BET).bet(1, 150).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'flop')).to.equal(false);
-  });
-
-  it('should allow to check multiple rounds.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 150).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 150).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'turn')).to.equal(false);
-  });
-
-  it('should allow to check multiple rounds, one left.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 150).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'turn')).to.equal(false);
-  });
-
-  it('should allow to check multiple rounds, all done.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_CHECK_TURN).checkTurn(1, 150).sign(P3_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 0, 'turn')).to.equal(true);
-  });
-
-  it('should not advance to next state if checks don\'t match state.', () => {
-    const lineup = [{
-      address: P1_ADDR,
-      last: new EWT(ABI_FOLD).fold(1, 5).sign(P1_PRIV),
-    }, {
-      address: P2_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 10).sign(P2_PRIV),
-    }, {
-      address: P3_ADDR,
-      last: new EWT(ABI_FOLD).fold(1, 5).sign(P3_PRIV),
-    }, {
-      address: P4_ADDR,
-      last: new EWT(ABI_CHECK_FLOP).checkFlop(1, 10).sign(P4_PRIV),
-    }];
-    expect(pokerHelper.isBettingDone(lineup, 1, 'turn')).to.equal(false);
-  });
 });
 
 describe('calculatePotsize() ', () => {
@@ -1339,16 +162,16 @@ describe('myMaxBet() ', () => {
     expect(pokerHelper.getMyMaxBet(lineup, P1_ADDR)).to.equal(200);
   });
 
-  it('should return -1 with no lineup.', () => {
-    expect(pokerHelper.getMyMaxBet(undefined, P1_ADDR)).to.equal(-1);
+  it('should throw with no lineup.', () => {
+    expect(pokerHelper.getMyMaxBet.bind(pokerHelper, undefined, P1_ADDR)).to.throw('invalid params.');
   });
 
-  it('should return -1 if not in lineup.', () => {
+  it('should throw if not in lineup.', () => {
     const lineup = [{
       address: P1_ADDR,
       last: new EWT(ABI_BET).bet(1, 200).sign(P1_PRIV),
     }];
-    expect(pokerHelper.getMyMaxBet(lineup, P2_ADDR)).to.equal(-1);
+    expect(pokerHelper.getMyMaxBet.bind(pokerHelper, lineup, P2_ADDR)).to.throw(`address ${P2_ADDR} not in lineup.`);
   });
 
   it('should return 0 if no receipt.', () => {
@@ -1429,28 +252,28 @@ describe('getSbPos() ', () => {
 
   it('should find SB Pos with empty seat at showdown.', () => {
     const lineup = [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 1000).sign(P1_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 1000).sign(P2_PRIV),
+      address: P1_ADDR,
+      last: new EWT(ABI_SHOW).show(1, 1000).sign(P1_PRIV),
+    }, {
+      address: P_EMPTY,
+    }, {
+      address: P2_ADDR,
+      last: new EWT(ABI_BET).bet(1, 1000).sign(P2_PRIV),
     }];
     expect(pokerHelper.getSbPos(lineup, 0, 'showdown')).to.equal(0);
   });
 
   it('should find SB Pos with all in at showdown.', () => {
     const lineup = [{
-        address: P1_ADDR,
-        last: new EWT(ABI_SHOW).show(1, 1000).sign(P1_PRIV),
-      }, {
-        address: P_EMPTY,
-      }, {
-        address: P2_ADDR,
-        last: new EWT(ABI_BET).bet(1, 1000).sign(P2_PRIV),
-        sitout: 'allin',
-      }];
+      address: P1_ADDR,
+      last: new EWT(ABI_SHOW).show(1, 1000).sign(P1_PRIV),
+    }, {
+      address: P_EMPTY,
+    }, {
+      address: P2_ADDR,
+      last: new EWT(ABI_BET).bet(1, 1000).sign(P2_PRIV),
+      sitout: 'allin',
+    }];
     expect(pokerHelper.getSbPos(lineup, 0, 'showdown')).to.equal(0);
   });
 });
@@ -1584,7 +407,7 @@ describe('findMaxRaise() ', () => {
     expect(pokerHelper.findMinRaiseAmount(lineup, 3, 100)).to.equal(100);
   });
 
-  it('should return -1 when there was no bet or raise', () => {
+  it('should throw when there was no bet or raise', () => {
     const lineup = [{
       address: P1_ADDR,
       last: new EWT(ABI_BET).bet(1, 100).sign(P1_PRIV),
@@ -1595,7 +418,7 @@ describe('findMaxRaise() ', () => {
       address: P3_ADDR,
       last: new EWT(ABI_BET).bet(1, 100).sign(P3_PRIV),
     }];
-    expect(pokerHelper.findMinRaiseAmount(lineup, 2, 100)).to.equal(-1);
+    expect(pokerHelper.findMinRaiseAmount.bind(pokerHelper, lineup, 2, 100)).to.throw('can not find minRaiseAmount.');
   });
 });
