+++ lib/receipt.js
@@ -12,7 +12,7 @@ export default class Receipt {
 
   // leave create a leave receipt
   static leave(...args) {
-    const [ tableAddr, handId, signerAddr ] = args;
+    const [tableAddr, handId, signerAddr] = args;
     // make leave receipt
       // <4 bytes hand ID>
       // <7 bytes table addr>
